+++ marginaleffects/__init__.py
@@ -1,5 +1,5 @@
 from .comparisons import avg_comparisons, comparisons
-from .datagrid import datagrid, datagridcf
+from .datagrid import datagrid
 from .hypotheses import hypotheses
 from .plot_comparisons import plot_comparisons
 from .plot_predictions import plot_predictions