+++ marginaleffects/model_pyfixest.py
@@ -32,7 +32,7 @@ class ModelPyfixest(ModelAbstract):
         if variables is None:
             variables = self.model._coefnames
             variables = [re.sub("\[.*\]", "", x) for x in variables]
-            variables = [x for x in variables if x in newdata.columns]
+            variables = [x for x in variables if x in self.modeldata.columns]
             variables = pl.Series(variables).unique().to_list()
         if isinstance(variables, (str, dict)):
             variables = [variables] if isinstance(variables, str) else variables